+++ changelog.md
@@ -1,3 +1,7 @@
+# Version 2024.12.7 (2024-12-20)
+
+- Rename create methods
+
 # Version 2024.12.7 (2024-12-18)
 
 - Extend element_matches_key search